add architecture targeting for armv8-a
Added -march=armv8-a to linuxarm64 compiler flags. This provides a small improvement in execution speed on Rockchip 3566 (and presumably 3588) by vectorising code using NEON. This benefit is approx 3.5% for rkmpp hardware decoding.
Armv8-a was chosen to limit the risk of breaking compatibility with other Arm64 targets,  e.g. if this is installed on a non-rk Arm64 board (35xx series are Armv8.2-a). There is no significant measured difference in decoding speed between an v8 and v8.2 target.